cavium: Always keep eMMC functional in non-secure world
While booting from MMC, SMMU was configured to treat eMMC transactions
as secure, thus eMMC wasn't functional in non-secure environment. Some
customers want to have access to MMC from NS-BL33 and because of that,
SMMU is restored to initial configuration.
It's necessary to note, that this solution is a hole in security system
and should be revisited.

Signed-off-by: Jan Dabros <jan.dabros@cavium.com>