trust: dts: Parse TRUST_ROT_BSSK property
TRUST_BSSK_ADDR property is set by BDK only if the HUK
is provided at build time. If not, TRUST_BSSK_ADDR is set
to 0.

As long as we use only SSK key at build time, the trust-key-addr
will be set to the FUSF_SSK base address. When ATF will have support
for HUK, FDT is already prepared for it.

Signed-off-by: Konrad Adamczyk <konrad.adamczyk@cavium.com>