psci: do not enable cache during cpu power-up
Currently, there's an open issue is some setups when
cpuidle is enabled - after sometime, one of the cores
might fail to come up after suspend.
This issue is under debug but generally it caused by
cache operations performed on the el3 stack.

For now, as a workaround, do not enable el3 cache,
in order to enable cpuidle feature.

Change-Id: Ic966046d238fb3d9be69394567938cd6339a4476
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/56075
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>
Tested-by: Kostya Porotchkin <kostap@marvell.com>