mvebu: ble: Make the BLE relocatable
The BLE (Binary extension to BootROM) is not a mainline ATF
component, but proprietary one for Marvell platforms.
This component should be relocatable and buildable as an
external component in order to be compatible with upstream
sources.
The patch leaves the "ble" folder in the current location
for compatibility with current CI builds, but relocates the
BLE build components from the main Makefile to platform make
files.
A new environemnt variable BLE_PATH should point to a location
of BLE folder and its default value is "ble".

Change-Id: I942da830029652bfb9c7a0ec2850e60059aa3126
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/57057
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>