fix: a8k-p: Change the way Makefile tests for PALLADIUM value
Change the A8K+ common makefile in the way that PALLADIUM variable
is always checked for value 1 and not 0, since PALLADIUM=0 is usually
not explicitly set in the compilation string.
Instead, the default value of 0 is assigned to PALLADIUM variable
by lately included make files in order to handle the preprocessor
directives in source files.

Change-Id: I9793e5808c56960a17cab0bab6ac4204fd3cbc27
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/58446
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>