mvebu: cp110: fix XFI link down problem
This patch fixes an issue on Armada7040/8040-DB,
when a XFI port is enabled, the link is not established.

This is caused due to changes in commit:
c14a9cf9535f ('commit mvebu: cp110: introduce porting layer for sfi').

This patch has the following updates:
- Fix incorrect static values in SFI configuration:
	g1_rx_selmuff, g1_rx_selmufi,
	g1_rx_selmupf, g1_rx_selmupi
- Update HPIPE_G1_SETTINGS_3_REG configuration.
- Remove the following settings:
	HPIPE_RX_CLK_ALIGN90_AND_TX_IDLE_CALIB_CTRL_REG
	HPIPE_DFE_REG0
	HPIPE_G1_SETTINGS_4_REG

In addition, cosmetic change in xfi_static_values_tab struct -
add 0x prefix for the values, to avoid inconsistency.

Change-Id: I97cb5f63504b6d16ccbc0b73c5ea037a83e8ba49
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/58930
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>