t9x: cgx: dt: configure CGX correctly when both QLM 3 & 7 are configured
Only QLM3 or QLM 7 can be configured for CGX0 at a time. But if a user
configures both QLM3 & 7, BDK fills up the SCRATCHX registers for both
QLM3 and QLM 7 but configures CGX0 for QLM7. Fixed parser to start
configuring CGX from higher QLM instead of 0 so QLM7 is configured for CGX0
instead of QLM3.

Signed-off-by: Chandrakala Chavva <cchavva@caviumnetworks.com>
Signed-off-by: Nithya <nithya.mani@caviumnetworks.com>