ecam: NDF: make NDF as secure device in t83, f95, t96
In T83XX, and onwards, nand flash controller is not available, though
it is found in PCI scanning. This makes the corresponding probe function
to be invoked when kernel boots up. Since the device is not present,
kernel shows traces of probe failure.

This patch marks this NFD as secure thus avoids PCI scanning of NDF
in kernel and thus avoids the error message.

Bug: OCTEONTX-1311

Signed-off-by: Sujeet Baranwal <sbaranwal@caviumnetworks.com>