octeontx2: flr: Add support for load/store register pair instructions
This patch add support for load and store rgister pair instructions.
SIMD instructions are excluded, because it will be hard to implement
and probably no one ever will use registers trapped by FLR in this
kind of instructions. Except that every variant of pair
store and load instruction is implemented.

Signed-off-by: Konrad Adamczyk <kadamczyk@marvell.com>