include: marvell: octeontx: Move helper functions
octeontx_build_mask(), octeontx_bit_extract(), octeontx_bit_insert()
and IS_OCTEONTX_PASS are simple inline functions or macros and are
moved to cavm_utils.h. This file was created for purpose to
contain simple helpers that can be useful accros whole OcteonTX
code and are not part of any specific module.

Change-Id: Id8576f418080acad391ede0f9e0fbc70c96497f1
Signed-off-by: Konrad Adamczyk <konrad.adamczyk@cavium.com>