plat: marvell: octeontx: platform_def.h: Fix BL1 and BL2 max size
TRUSTED_BOARD_BOOT is always defined, 1 for trusted boot and 0 for
non-trusted boot. #if defined(TRUSTED_BOARD_BOOT) is always true
and reduced BL1_RW_MAX_SIZE and BL2_RW_MAX_SIZE were never used.

This patch fix that issue. For T96 and F95 it was necessary to
increase size of RW memory for non-trusted image (which is still
smaller then for trusted image).

Change-Id: I053ea4b63572330ffc7722106caa4c8b405a111e
Signed-off-by: Tomasz Michalec <tomasz@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/7266
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <Chandrakala.Chavva@cavium.com>