drivers: marvell: phy: Fix serdes lane parameter in 6141 PHY API calls
Stop using the value in lmac_id as the serdes lane number in 6141 PHY API
calls because it's wrong for 50G (it only works for 10G and 25G).

Instead, get the lane number from the right-most two bits of
CGX(0..1)_CMR(0..3)_CONFIG[LANE_TO_SDS].  This works for 50G, 25G, and 10G.

Also, move a couple of ERROR() calls out of if/else statements to keep
lines shorter than 80 characters (to appease the checkpatch script).

Change-Id: I431765a6e1bad104e0c5c30564525887a6f75a85
Signed-off-by: Felix Manlunas <fmanlunas@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/7734
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>