plat: marvell: octeontx: otx2: loki: Fixes for running loki in ASIM
Because of skiping SCP in ASIM, scmi_init is skipped for now.
BDK assumes that when SCP is not working, then it is running on
emulator and eMMC boot device is selected. For now in ATF SPI0 CS0
is hardcoded as boot device.
Reading of GSERN_LANE_SCRATCH register is not working so part of
codes that use this register are skipped.

Change-Id: I0970791532ccd6f987c6146ad0d97a839105ec18
Signed-off-by: Tomasz Michalec <tomasz@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/10604
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>