drivers: marvell: cgx: Fix if link UP request not sent by u-boot
U-boot doesn't bring UP all the links by default as earlier.
Now, the links are brought UP based on the request by the user.
In that case, don't re-init the CGX link.

Also for speeds >= 10G and with PHY case, link status should
be read from CGX registers. If not, false link UP is reported.

Change-Id: I31d68912061265e81dda40f8b1458733933468c5
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/10617
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>