drivers: marvell: cgx: Fix LINK UP/DOWN issue
https://jira.cavium.com/browse/OCTEONTX-2085

Support CGX_CMD_LINK_BRING_UP and CGX_CMD_LINK_BRING_DOWN
at runtime in ATF. Though the framework was in place to
support these commands, it was not tested earlier. It is
supported now with this patch.

This patch also updates how the status info is populated
to the user for LINK UP command (also updates the doc),
    --- if the CGX configuration succeeds, then ATF returns
SUCCESS with whatever the link status is (Either UP/DOWN). It
is not treated as a error condition.
    --- if the CGX configuration fails, then ATF returns
FAILURE with respective error type set. No change in this method.

Change-Id: I3c6089f7b86eb1e70483518f1ca01478f5480e8c
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/7340
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>