drivers: marvell: cgx: Enable SFP management framework
This patch enables SFP management for the LMACs which has
sfp-slot or qsfp-slot info in DT (if updated by parser)

* In the LINK_UP command, if the LMAC has a SFP slot, it tries to read
the EEPROM capabilities, validate it and configure the CGX accordingly.
Error is thrown to user in case of mismatch between capabilities and
user options.
* Also during LINK_UP command, if the EEPROM data is not available
for some time or module is not present, the request is saved
to be processed later as an asynchronous event
* During poll timer #CB, if there is a change in link status
(disconnected->connected), and if the particular LMAC has SFP slot,
EEPROM capabilities are read again, capabilities are verified against
user options before re-configuring CGX.

Change-Id: I775bd500273b4aadd00bbf8cb373d11bf4d4d5b8
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/11797
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>