drivers: marvell: Add supported link modes
Supported LINK modes depends on each PHY capability
and based on SFP/QSFP slot. This patch adds API
to return the LINK modes for each PHY. some of them
may still be FIXME. this is a initial version.

This patch also updates SW FW data with the
supported link modes for each LMAC

Change-Id: Iae705c6f99aad18fca2c11c6996ee8b3b80da2cc
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/13279
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Nithya Mani <Nithya.Mani@cavium.com>