drivers: marvell: phy_mgmt: Use proper API when setting FEC type
When setting PHY modulation type to PAM4, FEC type must be set to RS-FEC
because it's required by PAM4.  But the code that does this neglects to use
the API for setting FEC type; rather it uses a direct assignment to
lmac->fec.  This causes a correct RS-FEC setting on the PHY, but it also
causes a cosmetic bug:  the "ethtool --show-fec" command shows a wrong
active FEC type; instead of showing RS-FEC (the current setting), it shows
the previous FEC setting.

Fix it by replacing the direct assignment to lmac->fec with a proper API
call.

Change-Id: I10ec5c82e38a7d350fc4fccd58f884365b328287
Signed-off-by: Felix Manlunas <fmanlunas@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/15105
Reviewed-by: Derek Chickles <Derek.Chickles@cavium.com>
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>