phy and cgx: Fix bugs in changing from NRZ to PAM4 and validating FEC
When RS-FEC is already on, changing PHY mod type from NRZ to PAM4 is only
partially successful:  PAM4 gets saved to SPI flash, but the PHY does not
switch to PAM4 unless there's a reboot (it ought to switch to PAM4
instantly).  The fix is to move the assignment:

    phy->mod_type = PHY_MOD_TYPE_PAM4;

from inside the if-statement (that checks the FEC type) to immediately
above the if-statement; this makes the assignment unconditional.  Also add
an else-clause that calls phy_config; this sends the actual instruction to
the PHY to switch to PAM4.

Also, change cgx_validate_fec_config() to take into account our proprietary
50GBASE-R4 PCS (used by v3 LIO3) in determining permitted FEC types.  That
particular PCS does not actually support FEC, but in LIO3, it is combined
with a PHY that uses standard 50GBASE-R2 and -R1 PCS (at the line-side),
which does support FEC.

Change-Id: Ia3ae784d687fc5883aa378d347e2a104e24689fb
Signed-off-by: Felix Manlunas <fmanlunas@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/15823
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Derek Chickles <Derek.Chickles@cavium.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>